// src/types/enums/pitch-deck-stage-param.enum.ts
// Enum to hold the query param values for the different stages
export enum PitchDeckStageParam {
    Problems = 'problems',
    Solutions = 'solutions',
    Product = 'product',
    IdeaValidation = 'idea-validation',
    CompetitorsAnalysis = 'competitors-analysis',
    BusinessModel = 'business-model',
    GoToMarketStrategy = 'go-to-market-strategy',
    UserAcquisitionStrategy = 'user-acquisition-strategy',
    Financials = 'financials',
    Roadmap = 'roadmap',
    Traction = 'traction',
    Team = 'team',
}